feat: unify overlay + coach into single dashboard
- Merge / and /coach into a single dashboard page (no longer transparent).
  Live match module sits on top, followed by last match insights, category
  rows, trend sparklines and a detailed today panel.
- Drop SPEED/BOOST live indicators and Avg boost / Avg supersonic from the
  today panel — already visible in the in-game HUD.
- Throttle UpdateState UI broadcasts from 15 fps to 5 s. Discrete events
  (BallHit, goals, shots) still propagate immediately.
- Replace the synthetic demo match loop with three hardcoded snapshot dicts
  published once at startup. /api/today and /api/coach also return them in
  demo mode so a refresh hydrates without waiting on the WS.
- /coach now responds with a 308 redirect to / for bookmark/back-compat.
- Self-reflection: kept fixture strings (player id/name) inline since they
  only appear inside a single preview-only block and are edited together.
