--- static/coach.html
@@ -1,93 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>RL Coach — stats to improve</title>
-    <meta name="description" content="Personal Rocket League dashboard: last match, positioning, boost and mechanics vs your own rolling average." />
-    <link rel="stylesheet" href="/static/coach.css" />
-  </head>
-  <body>
-    <header class="page-header">
-      <h1>
-        <span class="brand">COACH</span>
-        <span class="player" id="player-name">—</span>
-      </h1>
-      <label class="rank-picker">
-        <span class="visually-hidden">Target rank</span>
-        <select id="rank-select">
-          <option value="">No target</option>
-          <option value="bronze">Bronze</option>
-          <option value="silver">Silver</option>
-          <option value="gold">Gold</option>
-          <option value="platinum">Platinum</option>
-          <option value="diamond">Diamond</option>
-          <option value="champion">Champion</option>
-          <option value="grand_champion">Grand Champion</option>
-          <option value="supersonic_legend">Supersonic Legend</option>
-        </select>
-      </label>
-      <span class="conn" id="conn" aria-live="polite">connecting…</span>
-    </header>
-
-    <main class="page-main">
-      <section class="empty" id="empty" hidden>
-        <p>No matches recorded yet.</p>
-        <p class="hint">Start Rocket League and play a match. Come back here when you're done.</p>
-      </section>
-
-      <section class="hero" id="hero" hidden aria-labelledby="last-heading">
-        <header class="hero-head">
-          <h2 id="last-heading">Last match</h2>
-          <p class="hero-meta">
-            <span class="result" id="last-result">—</span>
-            <span class="sep" aria-hidden="true">·</span>
-            <span><strong id="last-score">0</strong> score</span>
-            <span class="sep" aria-hidden="true">·</span>
-            <span><strong id="last-duration">0</strong> min</span>
-          </p>
-        </header>
-        <ul class="insights" id="insights"></ul>
-        <p class="insights-empty" id="insights-empty" hidden>
-          Keep playing to surface patterns — I need at least 5 prior matches for suggestions.
-        </p>
-      </section>
-
-      <section class="cats" id="cats" hidden aria-label="Stat categories">
-        <article class="cat">
-          <h2>Positioning</h2>
-          <ul class="rows" id="cat-pos"></ul>
-        </article>
-        <article class="cat">
-          <h2>Boost (advanced)</h2>
-          <ul class="rows" id="cat-boost"></ul>
-        </article>
-        <article class="cat">
-          <h2>Mechanics</h2>
-          <ul class="rows" id="cat-mech"></ul>
-        </article>
-        <article class="cat">
-          <h2>Highlights</h2>
-          <ul class="rows" id="cat-highlights"></ul>
-        </article>
-      </section>
-
-      <section class="trend" id="trend-section" hidden aria-labelledby="trend-heading">
-        <h2 id="trend-heading">Trend <span class="muted">(last 14 days)</span></h2>
-        <ul class="trend-rows" id="trend-rows"></ul>
-      </section>
-
-      <section class="today-bar" id="today-bar" hidden aria-labelledby="today-heading">
-        <h2 id="today-heading" class="visually-hidden">Today</h2>
-        <span class="today-label">TODAY</span>
-        <span><strong id="today-matches">0</strong> matches</span>
-        <span><strong id="today-record">0W 0L</strong></span>
-        <span><strong id="today-winrate">0%</strong> WR</span>
-        <span><strong id="today-streak">0</strong> streak</span>
-        <span><strong id="today-best">0</strong> best score</span>
-      </section>
-    </main>
-
-    <script src="/static/coach.js" defer></script>
-  </body>
-</html>
